+++ proc_audio.py
@@ -11549,6 +11549,7 @@ class AudioProcessor:
             "spectral_flatness",
             "spectral_rolloff_hz_85",
             "spectral_rolloff_hz_95",
+            "roughness_parncutt_kernel",
             "roughness_aures_1985",
             "erb_weighted_spectral_density",
         ]